requirements.txt
flask
numpy
faiss-cpu
pillow
open-clip-torch